+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ dev
 - Fix methods.bo.utils.minimize to be strictly within bounds
 - Implemented the Two Stage Procedure, a method of summary-statistics diagnostics  
 - Added the MaxVar acquisition method
+- Added the RandMaxVar acquisition method
 
 0.6.3 (2017-09-28)
 ------------------